requirements.txt
absl-py==2.0.0
astunparse==1.6.3
bleach==6.1.0
blinker==1.4
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.0.3
colorama==0.4.4
command-not-found==0.3
contourpy==1.2.0
cryptography==3.4.8
cycler==0.12.1
dbus-python==1.2.18
distlib==0.3.7
distro==1.7.0
distro-info==1.1+ubuntu0.1
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.46.0
gast==0.5.4
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.2
h5py==3.10.0
httplib2==0.20.2
idna==3.4
imageio==2.33.1
importlib-metadata==4.6.4
jeepney==0.7.1
kaggle==1.5.16
keras==2.14.0
keyring==23.5.0
kiwisolver==1.4.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.3
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
ml-dtypes==0.2.0
more-itertools==8.10.0
netifaces==0.11.0
networkx==3.2.1
numpy==1.26.2
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvcc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.16.5
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
platformdirs==3.11.0
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydicom==2.4.3
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==5.4.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.22.0
scipy==1.11.4
SecretStorage==3.3.1
six==1.16.0
systemd-python==234
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.1
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
tensorrt==8.5.3.1
termcolor==2.3.0
text-unidecode==1.3
tifffile==2023.12.9
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.1.0
virtualenv==20.24.6
wadllib==1.3.6
webencodings==0.5.1
Werkzeug==3.0.1
wrapt==1.14.1
xkit==0.0.0
zipp==1.0.0
